+++ musify/src/views/LibraryView.vue
@@ -1,24 +1,12 @@
 <template>
   <v-container class="container">
+    <v-row> <h3 class="header">Albums</h3></v-row>
     <v-row class="row rounded-lg">
-      <h3 class="header">Albums</h3>
-      <v-carousel class="carousel" hide-delimiters>
-        <template :key="index" v-for="(albumChunk, index) in clearedAlbums">
-          <v-carousel-item>
-            <v-row class="carousel-row">
-              <template :key="i" v-for="(album, i) in albumChunk">
-                <v-col class="carousel-col">
-                  <AlbumItem :item="album"></AlbumItem>
-                </v-col>
-              </template>
-            </v-row>
-          </v-carousel-item>
-        </template>
-      </v-carousel>
+      <AlbumsList :albums="getAlbums"></AlbumsList>
     </v-row>
+    <v-row> <h3 class="header">Songs</h3></v-row>
     <v-row class="row rounded-lg">
       <section class="songs">
-        <h3 class="header">Songs</h3>
         <SongsList :songs="songs"></SongsList>
       </section>
     </v-row>
@@ -26,31 +14,23 @@
 </template>
 
 <script>
-import _ from "lodash";
-import AlbumItem from "@/components/AlbumItem";
 import SongsList from "@/components/SongsList";
+import AlbumsList from "@/components/AlbumsList";
 
 export default {
   name: "LibraryView",
   data() {
     return {
-      columns: 4,
       albums: [],
       songs: [],
     };
   },
 
   components: {
-    AlbumItem,
     SongsList,
+    AlbumsList,
   },
   computed: {
-    clearedAlbums() {
-      const filteredAlbums = _.uniqBy(this.albums, (album) => album.id);
-      const splitedAlbums = this.splitIntoChunks(filteredAlbums);
-      return splitedAlbums;
-    },
-
     getSongs() {
       return this.$store.getters.getSongs;
     },
@@ -59,34 +39,17 @@ export default {
       return this.$store.getters.getAlbums;
     },
   },
-  methods: {
-    splitIntoChunks(albums) {
-      const splitedAlbums = [];
-      let chunk = [];
-      let albumCount = 0;
+  methods: {},
 
-      for (let i = 0; i < albums.length; i++) {
-        chunk.push(albums[i]);
-        albumCount++;
-
-        if (albumCount === this.columns) {
-          splitedAlbums.push(chunk);
-          chunk = [];
-          albumCount = 0;
-        }
-      }
-      splitedAlbums.push(chunk);
-      chunk = 0;
-
-      return splitedAlbums;
+  watch: {
+    getAlbums: function (newAlbums) {
+      this.albums = newAlbums;
     },
   },
 
   mounted() {
-    this.$store.dispatch("fetchSongs", "eminem");
     this.albums = this.$store.getters.getAlbums;
     this.songs = this.$store.getters.getSongs;
-    console.log(this.songs);
   },
 };
 </script>
@@ -98,31 +61,14 @@ export default {
   gap: 20px;
 }
 
-.carousel-row {
-  height: 100%;
-  justify-content: center;
-}
-
 .row {
   box-shadow: 0 0 3px 1px rgba(34, 60, 80, 0.2);
-}
-
-.v-carousel {
-  height: 300px !important;
 }
 
 .header {
   color: #69686f;
   margin-left: 70px;
   margin-top: 20px;
-}
-
-.carousel-col {
-  flex-basis: 0;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  flex-grow: 0;
 }
 
 .songs {
